+++ .gitignore
@@ -95,3 +95,4 @@ ENV/
 
 # Specific Files
 settings.py
+sample/*
