+++ main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-from TwitchChatBotScrapper import TwitchChatBotScrapper
+from libs.TwitchChatBotScrapper import TwitchChatBotScrapper
 from settings import HOST, PORT, PASS, IDENT, CHANNEL
 
 chatBot = TwitchChatBotScrapper(HOST, PORT, PASS, IDENT, CHANNEL)
